chore : code changes

+++ scalar/src/lib.rs
@@ -14,11 +14,11 @@ pub fn pro(a: i8, b: i8) -> i8 {
 }
 
 #[allow(dead_code)]
-pub fn quo(a: f32, b: f32) -> f32 {
+pub fn quo(a: f64, b: f64) -> f64 {
     a / b
 }
 
 #[allow(dead_code)]
-pub fn rem(a: f32,b:f32)->f32 {
+pub fn rem(a: f64,b:f64)->f64 {
     a%b
 }
